+++ src/main/java/ch/reinhard/cashcontrol/modules/zahlung/domain/ZahlungEntityDetails.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @AllArgsConstructor
@@ -16,5 +17,5 @@ public class ZahlungEntityDetails {
     String empfaenger;
     String kategorieId;
     String text;
-    Long betrag;
+    BigDecimal betrag;
 }
