+++ src/main/java/ch/reinhard/cashcontrol/modules/zahlung/service/api/ZahlungDetailsDto.java
@@ -2,6 +2,7 @@ package ch.reinhard.cashcontrol.modules.zahlung.service.api;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Schema(name = "ZahlungDetails")
@@ -10,5 +11,5 @@ public record ZahlungDetailsDto(
     String empfaenger,
     String kategorieId,
     String text,
-    Long betrag
+    BigDecimal betrag
 ){}
