+++ src/main/java/ch/reinhard/cashcontrol/modules/zahlung/service/api/ZahlungDto.java
@@ -2,6 +2,7 @@ package ch.reinhard.cashcontrol.modules.zahlung.service.api;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Schema(name = "Zahlung")
@@ -12,5 +13,5 @@ public record ZahlungDto(
     String empfaenger,
     String kategorieId,
     String text,
-    Long betrag
+    BigDecimal betrag
 ){}
