+++ src/main/java/ch/reinhard/cashcontrol/modules/zahlung/service/api/ZahlungViewDto.java
@@ -2,6 +2,7 @@ package ch.reinhard.cashcontrol.modules.zahlung.service.api;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Schema(name = "ZahlungView")
@@ -13,6 +14,6 @@ public record ZahlungViewDto(
             String kategorieId,
             String kategorieBezeichnung,
             String text,
-            Long betrag
+            BigDecimal betrag
 ){}
